Refactoring code into a package module
