+++ src/lib.rs
@@ -4,6 +4,6 @@ extern crate serde;
 #[macro_use]
 extern crate serde_derive;
 
-mod error;
-mod carriers;
+pub mod error;
+pub mod packages;
 pub mod traits;
